orquestrador/requirements.txt
# orquestrador/requirements.txt
flask
requests
google-generativeai
# --- Novas Bibliotecas ---
langchain
langchain-google-genai
langchain-community